Update story.ts
Add story as prop, as it works as expected, but just isn't exposed.

I wanted to test out some functionality real quick and wanted to wrap the component in another tag to make sure some events worked correctly, but noticed this type was not exposed.

is there a reason for that? or just maybe overlooked?